+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * st-dropdown-menu: Allow user to move through options using the keyboard arrows 
 * st-select: new design for the search of elements.
+* st-search: Allow to hide or show loupe icon
 
 **Others:**